refactor(react): move the instance-element seam into its own module
The borrow host and its $$typeof install lived in component.ts / index.ts.
component.ts is loaded by the preact adapter via @expressive/react/adapter, so
preact bundled instance-element logic it never runs (it rejects direct instance
rendering) and reported component.ts short of full coverage.

Move the whole seam - the `$$typeof` brand and the `element` borrow host - into
element.ts, imported by the react entry as a side effect like ./jsx-runtime. The
brand's template is built lazily on first read, so the module can install before
the entry populates `Runtime`. The shared `frame` wrapper stays in component.ts,
since the ownership `render` host (which preact does use) needs it.
